предварительно описан кусок EventService

+++ ewm-stat/ewm-stat-dto/src/main/java/ru/practicum/ewm/dto/UpdateEventAdminRequest.java
@@ -30,7 +30,7 @@ public class UpdateEventAdminRequest { // Данные для изменения
 
     private Boolean paid; // новое значение флага о платности мероприятия
 
-    private Integer participantLimit; // новый лимит участников события
+    private Long participantLimit; // новый лимит участников события
 
     private Boolean requestModeration; // новое значение флага, нужна ли пре-модерация участников
